+++ source/extensions/omni.flux.utils.common/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [3.9.0]
+### Added
+- Added `elide_path` utility and `ElideModes` enum for truncating path strings with configurable ellipsis placement
+
 ## [3.8.2]
 ### Changed
 - Modernize python style and enable more ruff checks